Split profiles
