+++ spring-boot-2-demo-app/src/test/java/io/github/mfvanek/spring/boot2/test/support/RetryTestBase.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2020-2025. Ivan Vakhrushev and others.
+ * https://github.com/mfvanek/spring-boot-open-telemetry-demo
+ *
+ * Licensed under the Apache License 2.0
+ */
+
+package io.github.mfvanek.spring.boot2.test.support;
+
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test-retry")
+public abstract class RetryTestBase extends AbstractTestBase {
+
+}
