+++ spring-boot-3-demo-app/src/test/java/io/github/mfvanek/spring/boot3/test/service/PublicApiServiceTest.java
@@ -8,7 +8,7 @@
 package io.github.mfvanek.spring.boot3.test.service;
 
 import io.github.mfvanek.spring.boot3.test.service.dto.ParsedDateTime;
-import io.github.mfvanek.spring.boot3.test.support.TestBase;
+import io.github.mfvanek.spring.boot3.test.support.RetryTestBase;
 import io.micrometer.observation.Observation;
 import io.micrometer.observation.ObservationRegistry;
 import io.micrometer.tracing.Tracer;
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.system.CapturedOutput;
 import org.springframework.boot.test.system.OutputCaptureExtension;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
@@ -30,9 +29,8 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 import static com.github.tomakehurst.wiremock.client.WireMock.verify;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test-retry")
 @ExtendWith(OutputCaptureExtension.class)
-class PublicApiServiceTest extends TestBase {
+class PublicApiServiceTest extends RetryTestBase {
 
     @Autowired
     private PublicApiService publicApiService;
