+++ spring-boot-3-demo-app/src/test/java/io/github/mfvanek/spring/boot3/test/support/AbstractTestBase.java
@@ -1,0 +1,83 @@
+/*
+ * Copyright (c) 2020-2025. Ivan Vakhrushev and others.
+ * https://github.com/mfvanek/spring-boot-open-telemetry-demo
+ *
+ * Licensed under the Apache License 2.0
+ */
+
+package io.github.mfvanek.spring.boot3.test.support;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.tomakehurst.wiremock.client.WireMock;
+import io.github.mfvanek.spring.boot3.test.service.dto.CurrentTime;
+import io.github.mfvanek.spring.boot3.test.service.dto.ParsedDateTime;
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.actuate.observability.AutoConfigureObservability;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+import java.time.Clock;
+import java.util.TimeZone;
+import javax.annotation.Nonnull;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
+
+@AutoConfigureObservability
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ContextConfiguration(initializers = {KafkaInitializer.class, JaegerInitializer.class, PostgresInitializer.class})
+@AutoConfigureWireMock(port = 0)
+abstract class AbstractTestBase {
+
+    @Autowired
+    protected WebTestClient webTestClient;
+    @Autowired
+    protected ObjectMapper objectMapper;
+    @Autowired
+    protected Clock clock;
+
+    @BeforeEach
+    void resetExternalMocks() {
+        WireMock.resetAllRequests();
+    }
+
+    @Nonnull
+    protected String stubOkResponse(@Nonnull final ParsedDateTime parsedDateTime) {
+        final String zoneName = TimeZone.getDefault().getID();
+        stubOkResponse(zoneName, parsedDateTime);
+        return zoneName;
+    }
+
+    @SneakyThrows
+    private void stubOkResponse(@Nonnull final String zoneName, @Nonnull final ParsedDateTime parsedDateTime) {
+        final CurrentTime currentTime = new CurrentTime(parsedDateTime);
+        stubFor(get(urlPathMatching("/" + zoneName))
+            .willReturn(aResponse()
+                .withStatus(200)
+                .withBody(objectMapper.writeValueAsString(currentTime))
+            ));
+    }
+
+    @Nonnull
+    protected String stubErrorResponse() {
+        final String zoneName = TimeZone.getDefault().getID();
+        final RuntimeException exception = new RuntimeException("Retries exhausted");
+        stubErrorResponse(zoneName, exception);
+        return zoneName;
+    }
+
+    @SneakyThrows
+    private void stubErrorResponse(@Nonnull final String zoneName, @Nonnull final RuntimeException errorForResponse) {
+        stubFor(get(urlPathMatching("/" + zoneName))
+            .willReturn(aResponse()
+                .withStatus(500)
+                .withBody(objectMapper.writeValueAsString(errorForResponse))
+            ));
+    }
+}
